+++ src/app/admin/components/admin-product-create-page/admin-product-create-page.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-admin-product-create-page',
+  templateUrl: './admin-product-create-page.component.html',
+  styleUrls: ['./admin-product-create-page.component.scss']
+})
+export class AdminProductCreatePageComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}